Crate phc

RustCrypto: PHC String Format

Pure Rust implementation of the Password Hashing Competition (PHC) string format, which is used to store password hashes.

Example (Argon2):

$argon2d$v=19$m=512,t=3,p=2$5VtWOO3cGWYQHEMaYGbsfQ$AcmqasQgW/wI6wAHAMk4aQ

Minimum Supported Rust Version (MSRV) Policy

MSRV increases are not considered breaking changes and can happen in patch releases.

The crate MSRV accounts for all supported targets and crate feature combinations, excluding explicitly unstable features.

License

Licensed under either of:

Apache License, Version 2.0

MIT license

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Modules

error 🔒

Error types.

ident 🔒

Algorithm or parameter identifier.

output 🔒

Outputs from password hashing functions.

params 🔒

Algorithm parameters.

salt 🔒

Salt string support.

string_buf 🔒

value 🔒

Algorithm parameter value as defined by the [PHC string format].

Structs

Ident

Algorithm or parameter identifier.

Output

Output from password hashing functions, i.e. the “hash” or “digest” as raw bytes.

ParamsString

Algorithm parameter string.

PasswordHash

Password hash.

PasswordHashStringDeprecated

DEPRECATED: serialized PasswordHash.

Salt

In password hashing, a “salt” is an additional value used to personalize/tweak the output of a password hashing function for a given input password.

SaltString

“B64”-encoded Salt stored as a stack-allocated string.

Value

Algorithm parameter value string.

Enums

Error

Password hashing errors.

Constants

PASSWORD_HASH_SEPARATOR 🔒

Separator character used in password hashes (e.g. $6$...).

Type Aliases

Decimal

Type used to represent decimal (i.e. integer) values.

Result

Result type.